写好了性能和异常上报的工具类

+++ monitor/src/main/java/com/example/monitor/FPSCounter.java
@@ -36,7 +36,7 @@ public class FPSCounter {
         });
     }
 
-    public static double getFPS() {
-        return mFPS;
+    public static int getFPS() {
+        return (int) mFPS;
     }
 }
